+++ src/headless/converse-mam.js
@@ -23,7 +23,7 @@ const MAM_ATTRIBUTES = ['with', 'start', 'end'];
 
 converse.plugins.add('converse-mam', {
 
-    dependencies: ['converse-muc'],
+    dependencies: ['converse-disco', 'converse-muc'],
 
     overrides: {
         // Overrides mentioned here will be picked up by converse.js's
@@ -32,6 +32,64 @@ converse.plugins.add('converse-mam', {
         //
         // New functions which don't exist yet can also be added.
         ChatBox: {
+
+            fetchNewestMessages () {
+                /* Fetches messages that might have been archived *after*
+                 * the last archived message in our local cache.
+                 */
+                if (this.disable_mam) {
+                    return;
+                }
+                const { _converse } = this.__super__;
+                const most_recent_msg = u.getMostRecentMessage(this);
+
+                if (_.isNil(most_recent_msg)) {
+                    this.fetchArchivedMessages();
+                } else {
+                    const stanza_id = most_recent_msg.get(`stanza_id ${this.get('jid')}`);
+                    if (stanza_id) {
+                        this.fetchArchivedMessages({'after': stanza_id});
+                    } else {
+                        this.fetchArchivedMessages({'start': most_recent_msg.get('time')});
+                    }
+                }
+            },
+
+            async fetchArchivedMessages (options) {
+                if (this.disable_mam) {
+                    return;
+                }
+                const { _converse } = this.__super__;
+                const is_groupchat = this.get('type') === CHATROOMS_TYPE;
+                const mam_jid = is_groupchat ? this.get('jid') : _converse.bare_jid;
+                if (!(await _converse.api.disco.supports(Strophe.NS.MAM, mam_jid))) {
+                    return;
+                }
+                let message_handler;
+                if (is_groupchat) {
+                    message_handler = this.onMessage.bind(this);
+                } else {
+                    message_handler = _converse.chatboxes.onMessage.bind(_converse.chatboxes)
+                }
+                let result = {};
+                try {
+                    result = await _converse.api.archive.query(
+                        Object.assign({
+                            'groupchat': is_groupchat,
+                            'before': '', // Page backwards from the most recent message
+                            'max': _converse.archived_messages_page_size,
+                            'with': this.get('jid'),
+                        }, options));
+                } catch (e) {
+                    _converse.log(
+                        "Error or timeout while trying to fetch "+
+                        "archived messages", Strophe.LogLevel.ERROR);
+                    _converse.log(e, Strophe.LogLevel.ERROR);
+                }
+                if (result.messages) {
+                    result.messages.forEach(message_handler);
+                }
+            },
 
             async findDuplicateFromArchiveID (stanza) {
                 const { _converse } = this.__super__;
@@ -66,7 +124,26 @@ converse.plugins.add('converse-mam', {
                 }
                 return attrs;
             }
-        }
+        },
+
+        ChatRoom: {
+            initialize () {
+                this.__super__.initialize.apply(this, arguments);
+                this.on('change:mam_enabled', this.fetchArchivedMessagesIfNecessary, this);
+                this.on('change:connection_status', this.fetchArchivedMessagesIfNecessary, this);
+            },
+
+            fetchArchivedMessagesIfNecessary () {
+                if (this.get('connection_status') !== converse.ROOMSTATUS.ENTERED ||
+                        !this.get('mam_enabled') ||
+                        this.get('mam_initialized')) {
+                    return;
+                }
+                this.fetchArchivedMessages();
+                this.save({'mam_initialized': true});
+            }
+        },
+
     },
 
     initialize () {
@@ -139,6 +216,8 @@ converse.plugins.add('converse-mam', {
             }
         });
 
+        _converse.api.listen.on('afterMessagesFetched', chat => chat.fetchNewestMessages());
+        _converse.api.listen.on('chatReconnected', chat => chat.fetchNewestMessages());
         _converse.api.listen.on('addClientFeatures', () => _converse.api.disco.own.features.add(Strophe.NS.MAM));
         /************************ END Event Handlers ************************/